feat: 멤버 조회 기능 구현
