+++ src/main/resources/templates/admin/reservation-new.html
@@ -101,12 +101,8 @@
 
 <script src="/js/user-scripts.js"></script>
 
-<script src="/js/reservation-new.js"></script>
+<!--<script src="/js/reservation-new.js"></script>-->
 
-<!--
-  TODO: [5단계] 예약 생성 기능 변경 - 관리자
-        reservation-new.js 파일 대신 reservation-with-member.js 파일 사용
--->
-<!--<script src="/js/reservation-with-member.js"></script>-->
+<script src="/js/reservation-with-member.js"></script>
 </body>
 </html>
