+++ src/test/java/roomescape/controller/LoginControllerTest.java
@@ -21,7 +21,7 @@ import static org.hamcrest.Matchers.is;
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 @DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
 @Sql(scripts = {"/test_schema.sql", "/test_data.sql"})
-public class MemberControllerTest {
+public class LoginControllerTest {
 
     @LocalServerPort
     int port;
